refactor(env.py, machine.py, plugins): improve plugin architecture and remove redundant modules
Refactor the plugin architecture to enhance modularity and maintainability. Introduce new typing and return types for better type safety and clarity. Remove redundant plugin modules (nvim, powershell, python, tailscale, vscode, zed) and consolidate their functionalities into a more streamlined structure. This reduces code duplication and simplifies the plugin management process. The changes also include the addition of new configuration and environment protocols for better configuration management. The refactoring aims to make the codebase more scalable and easier to extend with new plugins in the future.
