+++ app/plugins/__init__.py
@@ -1,12 +1,7 @@
 """Individual machine components setup package."""
 
 from .git import *
-from .nvim import *
 from .private_files import *
-from .python import *
 from .shell import *
 from .ssh import *
-from .tailscale import *
 from .tools import *
-from .vscode import *
-from .zed import *
